refactor: getImageByUrlをどこでも呼べるように切り出し
